+++ NEWS.md
@@ -1,6 +1,6 @@
 # bstfun (development version)
 
-* Updated `auc_density()` to better handle cut points and 0 and 1.
+* Updated `auc_density()` to better handle cut points and 0 and 1 and improved output structure by adding sensitivity and specificity estimates to a tibble.
 
 * Added new functions `add_n.tbl_likert()` and `add_continuous_stat().tbl_likert()` (#55)